mob/requirements.txt
# Azure AI Search Dependencies
azure-search-documents>=11.4.0
azure-core>=1.29.0

# OpenAI for embeddings
openai>=1.0.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0

# File Processing (for real implementation)
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.0.9
python-pptx>=0.6.21

# Language Processing
langdetect>=1.0.9

# Utilities
python-dateutil>=2.8.2
typing-extensions>=4.0.0
python-dotenv>=1.0.0 